well, I'm redoing a lot of this.
I figured out the dropup but I really don't like the way it looks. I also learned (thanks bootstrap docs) that dropdown menus can have forms in them.

The new plan given the above information is to create a template for the boxes that the character sheets will go in and include a + dropdown button that opens a form that asks for info to create the character. From there, take that info and reprocess the current box to instead have a charcter with that info, then create a new box with the plus in it. I have no clue how this will all end up working. It sounds like html recursion but I really don't know. This is the fun of research, I guess.