DB is organized and stable - basically what the title says. Most of these changes were to the model. - added json support for all tables - added relationships between all tables and the character table - added methods to serialize an individual and the whole party
This all results in very nice json that is being communicated with the frontend. Now all I need to do is use Jinja to template index based on that json and I actually have a website (or at least a skeleton)

One problem is that I need to redraw the entire page on each update. This should be okay short term, but eventually there will be some js/jquery action to make sure that the minimum viable amount of data is being sent on each update.
Another problem is carry weight - I think I will need to use JS to accurately calc/update the total, but for now I am hoping jinja will let me store a template to calc it. My second best idea is just updating the stats entry for a character every time an item is added to (or removed from) their inventory